Added a possibility to pick a version of utils on travis

+++ ci-integration/travis.sh
@@ -1,7 +1,8 @@
 #!/bin/bash
 
+echo " >> Installing RiotKit CI-Utils on this Travis-CI build"
 mkdir -p /opt/riotkit/utils
-wget https://github.com/riotkit-org/ci-utils/archive/master.zip -O /tmp/ci-utils.zip
+test -f /tmp/ci-utils.zip || wget https://github.com/riotkit-org/ci-utils/archive/master.zip -O /tmp/ci-utils.zip
 unzip /tmp/ci-utils.zip -d /opt/riotkit/utils
 mv /opt/riotkit/utils/ci-utils*/* /opt/riotkit/utils/
 rm -rf /opt/riotkit/utils/ci-utils*/
